backend/requirements.txt
pandas
mapillary
sqlalchemy
fastapi
psycopg
matplotlib
imagehash
Pillow
aiohttp
tqdm